+++ R/findLocation.R
@@ -1,6 +1,6 @@
 #' Find host-parasite interactions for a given location
 #'
-#' Given a location (available from \code{\link{listLocations}} this function 
+#' Given a location (available from \code{data{locations}) this function 
 #'		returns all host-parasite associations in that location.
 #'
 #'
@@ -49,11 +49,13 @@ findLocation <- function(location = NULL, group = NULL, citation = FALSE,
 	hostState = NULL, speciesOnly = FALSE, 
 	validateHosts = FALSE, removeDuplicates=FALSE){
   if(is.null(location)){
-    stop("Please choose a location from the possible locations in the listLocations() function")
+        stop("Please choose a location from the possible 
+					locations... data(locations)")
   }
 
   if (location %in% locations[,1] == FALSE) {
-    stop("Please choose a location from the possible locations in the listLocations() function")
+        stop("Please choose a location from the possible 
+					locations... data(locations)")
     }
 
 #    location1 <- gsub("\\+", "%2B", location)