+++ R/helminthR-package.R
@@ -11,12 +11,12 @@
 #' <http://www.nhm.ac.uk/research-curation/scientific-resources/taxonomy-systematics/host-parasites/>
 #'
 #' @importFrom utils tail globalVariables
-#' @importFrom ggmap geocode
 #' @importFrom rvest html html_nodes html_text html_attr
 #' @importFrom magrittr %>%
 #' @importFrom plyr ldply
 #' @importFrom xml2 read_html read_xml xml_attr xml_text xml_children xml_find_all
 #' @importFrom httr content GET stop_for_status
+#' @importFrom utils data
 #' @name helminthR-package
 #' @aliases helminthR
 #' @docType package