+++ R/locations.R
@@ -1,6 +1,13 @@
 #' Table of geographic location names, and associated coordinates
 #'
-#' A cached copy of the output of \code{\link{listLocations}}.
+#' Lists geographic locations that can be input to \code{\link{findHost}} or
+#' \code{\link{findParasite}} and the corresponding latitude and longitude coordinates 
+#' of the country's centroid. The georeferencing was performed dynamically using the 
+#' Google Maps API, but they have since restricted access. The data on locations is now
+#' provided in this data file called `locations` -- `data(locations)` -- and is based on 
+#' an earlier usage of \code{ggmap}. The geographic coordinates may not be accurate, and users
+#' should check for accuracy (and feel free to file an issue or PR on Github with corrections).
+#' 
 #'
 #' @format
 #'  \describe{